foodagent/requirements.txt
absl-py==2.1.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
astunparse==1.6.3
attrs==25.3.0
backoff==2.2.1
bcrypt==4.3.0
bidict==0.23.1
biosppy==2.2.2
build==1.2.2.post1
cachetools==5.5.2
certifi==2024.8.30
cffi==2.0.0
charset-normalizer==3.4.0
chromadb==1.0.9
click==8.1.8
coloredlogs==15.0.1
contourpy==1.3.1
cryptography==46.0.3
cycler==0.12.1
Deprecated==1.2.18
distlib==0.3.9
distro==1.9.0
Django==5.1.6
django-allauth==65.13.0
django-anymail==12.0
django-appconf==1.1.0
django-compressor==4.5.1
django-leaflet==0.31.0
django-sass-processor==1.4.1
django_components==0.130
djangorestframework==3.16.1
djc_core_html_parser==1.0.1
durationpy==0.9
fastapi==0.115.9
filelock==3.18.0
flatbuffers==24.3.25
fontawesomefree==6.6.0
fonttools==4.54.1
fsspec==2025.3.2
gast==0.6.0
GDAL==3.9.2
google-auth==2.40.1
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.67.1
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.8
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.31.2
humanfriendly==10.0
idna==3.10
importlib_metadata==8.6.1
importlib_resources==6.5.2
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
jwt==1.4.0
keras==3.6.0
kiwisolver==1.4.7
kubernetes==32.0.1
langchain==1.0.3
langchain-core==1.0.2
langgraph==1.0.2
langgraph-checkpoint==3.0.0
langgraph-prebuilt==1.0.2
langgraph-sdk==0.2.9
langsmith==0.4.39
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.9.2
mdurl==0.1.2
ml-dtypes==0.4.1
mmh3==5.1.0
mock==5.1.0
mpmath==1.3.0
namex==0.0.8
numpy==2.0.2
oauthlib==3.2.2
onnxruntime==1.22.0
openai==1.75.0
opencv-python==4.10.0.84
opentelemetry-api==1.33.0
opentelemetry-exporter-otlp-proto-common==1.33.0
opentelemetry-exporter-otlp-proto-grpc==1.33.0
opentelemetry-instrumentation==0.54b0
opentelemetry-instrumentation-asgi==0.54b0
opentelemetry-instrumentation-fastapi==0.54b0
opentelemetry-proto==1.33.0
opentelemetry-sdk==1.33.0
opentelemetry-semantic-conventions==0.54b0
opentelemetry-util-http==0.54b0
opt_einsum==3.4.0
optree==0.13.1
orjson==3.10.18
ormsgpack==1.11.0
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
platformdirs==4.3.8
posthog==4.0.1
protobuf==5.28.3
psycopg2==2.9.10
psycopg2-binary==2.9.11
py4j==0.10.9.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
Pygments==2.18.0
PyMySQL==1.1.1
pyparsing==3.2.0
PyPika==0.48.9
pypng==0.20220715.0
pyproject_hooks==1.2.0
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2024.2
PyWavelets==1.7.0
PyYAML==6.0.2
qrcode==7.4.2
rcssmin==1.1.2
referencing==0.36.2
regex==2025.10.23
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rich==13.9.4
rjsmin==1.2.2
rpds-py==0.25.0
rsa==4.9.1
scikit-learn==1.5.2
scipy==1.14.1
setuptools==75.4.0
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
sniffio==1.3.1
soundfile==0.12.1
sqlparse==0.5.3
starlette==0.45.3
stripe==13.1.1
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.18.0
tensorboard-data-server==0.7.2
tensorflow==2.18.0
termcolor==2.5.0
threadpoolctl==3.5.0
tiktoken==0.12.0
tokenizers==0.21.1
tqdm==4.67.1
typer==0.15.4
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.34.1
uvloop==0.21.0
virtualenv==20.31.2
watchfiles==1.0.5
websocket-client==1.8.0
websockets==15.0.1
Werkzeug==3.1.3
wfdb==4.1.2
wheel==0.45.0
whitenoise==6.9.0
wrapt==1.16.0
xxhash==3.6.0
zipp==3.21.0
zstandard==0.25.0
langchain_openai